requirements.txt
-e git://github.com/jaywink/aiolog.git@matrix-handler#egg=aiolog
aiohttp
email-validator>=1.1.2
matrix-nio[e2e]>=0.18.0
Markdown>=3.1.1
python-keycloak>=0.24.0
PyYAML>=5.1.2
requests
